# openapi/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "additionalProperties": false,
        "description": "The single error shape returned by every Acme Commerce API error response.",
        "example": {
          "error": {
            "code": "VALIDATION_ERROR",
            "details": {
              "fields": [
                {
                  "field": "body.title",
                  "message": "\"title\" is required.",
                  "rule": "required"
                }
              ]
            },
            "message": "The request body is invalid.",
            "request_id": "req_18f3a9c4e21b7d05f6a3b8c1"
          }
        },
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable machine-readable error code. Branch on this, never on `message`. The full list is documented in README.md.",
                "example": "SKU_ALREADY_EXISTS",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "description": "Optional structured context. Its shape depends on `code`: `VALIDATION_ERROR` carries a `fields` array, `MALFORMED_ID` carries `expected_pattern`, `SKU_ALREADY_EXISTS` carries the conflicting `sku`.",
                "properties": {},
                "type": "object"
              },
              "message": {
                "description": "Human-readable explanation, for a log or a response pane. May be reworded in any release; it is not part of the contract.",
                "type": "string"
              },
              "request_id": {
                "description": "Correlation identifier for this request, also returned in the `X-Request-Id` response header. Quote it when reporting a problem — it is what locates the server-side log lines.",
                "example": "req_18f3a9c4e21b7d05f6a3b8c1",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "Location": {
        "additionalProperties": false,
        "description": "A place where inventory is held.",
        "properties": {
          "address_line1": {
            "description": "Street address. Null for a virtual location.",
            "maxLength": 200,
            "type": [
              "null",
              "string"
            ]
          },
          "address_line2": {
            "description": "Unit, suite, or floor.",
            "maxLength": 200,
            "type": [
              "null",
              "string"
            ]
          },
          "city": {
            "description": "City.",
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "description": "ISO 3166-1 alpha-2 country code, uppercase.",
            "maxLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "created_at": {
            "description": "RFC 3339 creation timestamp.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Server-assigned identifier.",
            "pattern": "^loc_[0-9a-f]{24}$",
            "type": "string"
          },
          "is_active": {
            "description": "Whether the location is in use. Locations are never deleted — inventory history references them — so retiring one means setting this to false.",
            "type": "boolean"
          },
          "name": {
            "description": "Human-readable name. Unique, case-insensitively.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "postal_code": {
            "description": "Postal or ZIP code.",
            "maxLength": 20,
            "type": [
              "null",
              "string"
            ]
          },
          "region": {
            "description": "Province, state, or region.",
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "description": "Kind of location. `warehouse` fulfils orders, `retail` is a physical store, `virtual` covers dropship and in-transit stock that has no address.",
            "enum": [
              "warehouse",
              "retail",
              "virtual"
            ],
            "type": "string"
          },
          "updated_at": {
            "description": "RFC 3339 last-modified timestamp.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "address_line1",
          "address_line2",
          "city",
          "region",
          "postal_code",
          "country",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "Location",
        "type": "object"
      },
      "LocationCreate": {
        "additionalProperties": false,
        "description": "Body for creating an inventory location.",
        "example": {
          "address_line1": "3020 Centre Street NE",
          "city": "Calgary",
          "country": "CA",
          "name": "Calgary Warehouse",
          "postal_code": "T2E 2X7",
          "region": "AB",
          "type": "warehouse"
        },
        "properties": {
          "address_line1": {
            "description": "Street address.",
            "maxLength": 200,
            "type": "string"
          },
          "address_line2": {
            "description": "Unit or suite.",
            "maxLength": 200,
            "type": "string"
          },
          "city": {
            "description": "City.",
            "maxLength": 100,
            "type": "string"
          },
          "country": {
            "description": "ISO 3166-1 alpha-2, uppercase.",
            "pattern": "^[A-Z]{2}$",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "description": "Defaults to true.",
            "type": "boolean"
          },
          "name": {
            "description": "Required. Must be unique.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "postal_code": {
            "description": "Postal or ZIP code.",
            "maxLength": 20,
            "type": "string"
          },
          "region": {
            "description": "Province or state.",
            "maxLength": 100,
            "type": "string"
          },
          "type": {
            "description": "Kind of location. `warehouse` fulfils orders, `retail` is a physical store, `virtual` covers dropship and in-transit stock that has no address.",
            "enum": [
              "warehouse",
              "retail",
              "virtual"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "LocationCreate",
        "type": "object"
      },
      "LocationListResponse": {
        "additionalProperties": false,
        "description": "A page of locations with pagination metadata.",
        "properties": {
          "data": {
            "description": "The requested page of locations.",
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "LocationListResponse",
        "type": "object"
      },
      "LocationResponse": {
        "additionalProperties": false,
        "description": "A single location.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "required": [
          "data"
        ],
        "title": "LocationResponse",
        "type": "object"
      },
      "LocationUpdate": {
        "additionalProperties": false,
        "description": "Partial update. Omitted properties are unchanged; explicit null clears one.",
        "example": {
          "is_active": false
        },
        "minProperties": 1,
        "properties": {
          "address_line1": {
            "description": "New street address; null clears it.",
            "maxLength": 200,
            "type": [
              "string",
              "null"
            ]
          },
          "address_line2": {
            "description": "New unit or suite; null clears it.",
            "maxLength": 200,
            "type": [
              "string",
              "null"
            ]
          },
          "city": {
            "description": "New city; null clears it.",
            "maxLength": 100,
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "description": "New ISO 3166-1 alpha-2 code; null clears it.",
            "pattern": "^[A-Z]{2}$",
            "type": [
              "string",
              "null"
            ]
          },
          "is_active": {
            "description": "Set false to retire the location without deleting it.",
            "type": "boolean"
          },
          "name": {
            "description": "New name.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "postal_code": {
            "description": "New postal code; null clears it.",
            "maxLength": 20,
            "type": [
              "string",
              "null"
            ]
          },
          "region": {
            "description": "New region; null clears it.",
            "maxLength": 100,
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "description": "Kind of location. `warehouse` fulfils orders, `retail` is a physical store, `virtual` covers dropship and in-transit stock that has no address.",
            "enum": [
              "warehouse",
              "retail",
              "virtual"
            ],
            "type": "string"
          }
        },
        "title": "LocationUpdate",
        "type": "object"
      },
      "Pagination": {
        "additionalProperties": false,
        "description": "Offset pagination metadata, present on every collection response.",
        "properties": {
          "limit": {
            "description": "The page size that was applied.",
            "minimum": 1,
            "type": "integer"
          },
          "page": {
            "description": "The page that was returned.",
            "minimum": 1,
            "type": "integer"
          },
          "total": {
            "description": "Total records matching the filters, across all pages.",
            "minimum": 0,
            "type": "integer"
          },
          "total_pages": {
            "description": "Number of pages at this `limit`. Zero when `total` is zero — not one empty page.",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "total_pages"
        ],
        "title": "Pagination",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "A development bearer token from `POST /api/v1/auth/token`.\n\nSend it as `Authorization: Bearer <token>`.\n\nThe token is a JWT: signed, **not encrypted**. Anyone holding it can read every claim inside it, so treat it as a credential and keep it out of anything you commit. It cannot be revoked individually — it is valid until it expires.\n\nCatalog **reads** are deliberately unauthenticated: a storefront browses the catalog without credentials. Everything that writes, and everything under locations, inventory, and pricing, requires a token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
  "info": {
    "contact": {
      "email": "platform@acme.example",
      "name": "Acme Commerce Platform Team"
    },
...
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Internal ecommerce platform API for catalog, inventory, orders, and fulfillment.",
    "title": "Acme Commerce API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/locations": {
      "get": {
        "description": "Paginated, filterable list of the places inventory is held.\n\nWith no `is_active` filter, both active and retired locations are returned — the same\nno-hidden-filter rule as the catalog. The result set is fully determined by the query\nstring.\n\nRequires the `locations:read` permission, which every role has.",
        "operationId": "listLocations",
        "parameters": [
          {
            "description": "One-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Records per page, max 100.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Kind of location. `warehouse` fulfils orders, `retail` is a physical store, `virtual` covers dropship and in-transit stock that has no address.",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "enum": [
                "warehouse",
                "retail",
                "virtual"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by active state. Omit to return locations of both states.",
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Case-insensitive substring match on name and city.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Sort field. Allowlisted; `id` is appended as a stable tiebreaker.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name",
              "enum": [
                "name",
                "type",
                "created_at",
                "updated_at"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort direction.",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationListResponse"
                }
              }
            },
            "description": "A page of locations."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is malformed (`VALIDATION_ERROR`, `MALFORMED_ID`, or `INVALID_JSON`)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing, malformed, invalid, or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated, but your role lacks the required permission (`INSUFFICIENT_PERMISSION`)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The database is unreachable. Safe to retry."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List inventory locations",
        "tags": [
          "Locations"
        ]
      },
      "post": {
        "description": "Creates a location. Names are unique case-insensitively; a duplicate returns `409`.\n\nRequires `locations:write`. The `support` role does not have it — logging in as\nsupport and calling this is the shortest way to see a `403` that names the permission\nit wanted.",
        "operationId": "createLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            },
            "description": "The location that was created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is malformed (`VALIDATION_ERROR`, `MALFORMED_ID`, or `INVALID_JSON`)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing, malformed, invalid, or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated, but your role lacks the required permission (`INSUFFICIENT_PERMISSION`)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A location with that name already exists (`LOCATION_NAME_EXISTS`)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The database is unreachable. Safe to retry."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an inventory location",
        "tags": [
          "Locations"
        ]
      }
    },
    "/api/v1/locations/{locationId}": {
      "get": {
        "description": "Returns one location. A syntactically invalid id returns `400 MALFORMED_ID`; a well-formed id naming nothing returns `404`.",
        "operationId": "getLocation",
        "parameters": [
          {
            "description": "Location identifier as returned by the API.",
            "in": "path",
            "name": "locationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            },
            "description": "The requested location."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is malformed (`VALIDATION_ERROR`, `MALFORMED_ID`, or `INVALID_JSON`)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing, malformed, invalid, or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated, but your role lacks the required permission (`INSUFFICIENT_PERMISSION`)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No location exists with that identifier (`LOCATION_NOT_FOUND`)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The database is unreachable. Safe to retry."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Retrieve a location",
        "tags": [
          "Locations"
        ]
      },
      "patch": {
        "description": "Updates only the properties present in the body. Omitted properties are unchanged; an\nexplicit `null` clears a nullable one.\n\n**There is no DELETE.** A location is referenced by inventory levels and by the\nadjustment audit log, so removing one would orphan history that has to stay\nanswerable. Retire a location with `{\"is_active\": false}` instead — which is what\n\"delete this warehouse\" actually means in a business that has shipped from it.",
        "operationId": "updateLocation",
        "parameters": [
          {
            "description": "Location identifier as returned by the API.",
            "in": "path",
            "name": "locationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            },
            "description": "The location after the update."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is malformed (`VALIDATION_ERROR`, `MALFORMED_ID`, or `INVALID_JSON`)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing, malformed, invalid, or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated, but your role lacks the required permission (`INSUFFICIENT_PERMISSION`)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No location exists with that identifier (`LOCATION_NOT_FOUND`)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Another location already uses that name (`LOCATION_NAME_EXISTS`)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The database is unreachable. Safe to retry."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Partially update a location",
        "tags": [
          "Locations"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "The server this document was fetched from.",
      "url": "/"
    }
  ],
...
}